Corregido error en la llamada a los Prodcutos. Igual que en el commit anterior se ha cambiado el tipo de dato a decimal y se ha modificado la carga de la connectioString que estaba errónea

+++ Models/Producto.cs
@@ -4,7 +4,7 @@ public class Producto{
 
     public int? idProducto {get; set;}
     public string? descripcion {get; set;}
-    public double? precio {get; set;}
+    public decimal? precio {get; set;}
     public int? idTipoIVA {get; set;}
     public DateTime? fechaCreacion  {get; set;}
     public bool activo {get; set;}
@@ -15,7 +15,7 @@ public class Producto{
         activo = true;
     }
 
-    public Producto(int _idProducto, string _descripcion, double _precio, int _idTipoIVA, DateTime _fechaCreacion)
+    public Producto(int _idProducto, string _descripcion, decimal _precio, int _idTipoIVA, DateTime _fechaCreacion)
     {
         idProducto = _idProducto;
         descripcion = _descripcion;
